Catalog Fixes
Started rework on database by adding a map for table id and number of pages in a table to the catalog. This will replace having to keep the number of pages and the page order in the header of a table file. This will allow for no longer needing to rewrite the table header and use page order.